redesign of file loading
the old file loading mechanism planted the text of the submitted
file directly into an html textarea, leading to uncontrollable
XSS vulnerabilities. The new mechanism loads the text using a
distinct ajax call, and passes the string as a string directly
to a set-contents method of CodeMirror. The Browse page is now
working with the new mechanism. The non-browse file container
needs to be updated as well. This was a big change.

Note that this change goes hand in hand with changes to the js/ts
in the admiral-edu server, specifically commit bbd3006b1b58
(of course, if I rebase, that hash may disappear....)